refactor: minor package changes, added `SquareTextView` in demo
