+++ demo/src/main/java/com/subinkrishna/demo/ui/ViewItem.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright (C) 2018 Subinkrishna Gopi
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.subinkrishna.demo.ui
+
+/**
+ * @author Subinkrishna Gopi
+ */
+sealed class ViewItem {
+    data class Image(val url: String, val ratio: Float) : ViewItem()
+    data class Text(val text: String, val ratio: Float = 1f) : ViewItem()
+}
