+++ demo/src/main/res/layout/view_text_item.xml
@@ -1,12 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.subinkrishna.aspect.AspectRatioTextView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<com.subinkrishna.demo.ui.SquareTextView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="#0277bd"
     android:gravity="bottom|start"
     android:padding="10dp"
     android:textColor="#fff"
-    android:textSize="14sp"
-    app:lock="width"
-    app:ratio="1" />
+    android:textSize="14sp" />
